backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Cache and Queue
redis==5.0.1
celery==5.3.4

# Authentication and Security
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
python-multipart==0.0.6
Authlib==1.6.4
limits==5.5.0

# Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-ulid==2.7.0
python-dotenv==1.0.0
httpx==0.25.2

# File Processing and PDF
beautifulsoup4==4.12.2
lxml==4.9.3
weasyprint==60.2
reportlab==4.0.7

# Storage
minio==7.2.0

# Rate Limiting
slowapi==0.1.9

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
ruff==0.1.6
